feat: add dynamic temp and build directories
This commit introduces dynamic temporary and build directories for the
rofi-calc role. The directories are now defined using variables, which
improves flexibility and maintainability. The hardcoded paths have been
replaced with the variables `tmp_dir` and `build_dir`.

Additionally, the `vars/main.yml` file has been deleted as it is no
longer needed.